ADD - 식약처 레시피 정보 Import Parser, 코드 순서 변경

+++ api/src/main/java/com/ongi/api/admin/web/dto/ParsedIngredient.java
@@ -4,7 +4,7 @@ import com.ongi.ingredients.domain.enums.RecipeIngredientUnitEnum;
 
 public record ParsedIngredient(
 	String name, // 재료명 (예: 연두부)
-	Integer quantity, // 수량 (예: 75)
+	Double quantity, // 수량 (예: 75)
 	RecipeIngredientUnitEnum unit, // 단위 (예: G)
 	String note // 비고 (예: 3/4모, 다진 것 등)
 ) { }
